backend/requirements.txt
fastapi
uvicorn
python-dotenv
numpy
sentence-transformers
psycopg2-binary
sqlalchemy
pgvector
requests